refactor: useCommentList 커스텀 훅 작성
